+++ tests/test_checker/test_noqa.py
@@ -95,6 +95,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z323': 0,
         'Z324': 1,
         'Z325': 1,
+        'Z326': 1,
 
         'Z400': 0,
         'Z401': 0,