+++ tests/test_visitors/test_tokenize/test_primitives/test_string_tokens/test_unicode_prefix.py
@@ -4,9 +4,10 @@ import pytest
 
 from wemake_python_styleguide.violations.consistency import (
     UnicodeStringViolation,
+    UppercaseStringModifierViolation,
 )
 from wemake_python_styleguide.visitors.tokenize.primitives import (
-    WrongPrimitivesVisitor,
+    WrongStringTokenVisitor,
 )
 
 
@@ -28,7 +29,7 @@ def test_unicode_prefix(
     """Ensures that unicode prefixes raise a warning."""
     file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
-    visitor = WrongPrimitivesVisitor(default_options, file_tokens=file_tokens)
+    visitor = WrongStringTokenVisitor(default_options, file_tokens=file_tokens)
     visitor.run()
 
     assert_errors(visitor, [UnicodeStringViolation])
@@ -54,7 +55,26 @@ def test_correct_strings(
     """Ensures that correct strings are fine."""
     file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
-    visitor = WrongPrimitivesVisitor(default_options, file_tokens=file_tokens)
+    visitor = WrongStringTokenVisitor(default_options, file_tokens=file_tokens)
     visitor.run()
 
     assert_errors(visitor, [])
+
+
+def test_unicode_regression(
+    parse_tokens,
+    assert_errors,
+    default_options,
+    primitives_usages,
+    mode,
+):
+    """Ensures that correct uppercase unicode string raises two violations."""
+    file_tokens = parse_tokens(mode(primitives_usages.format('U""')))
+
+    visitor = WrongStringTokenVisitor(default_options, file_tokens=file_tokens)
+    visitor.run()
+
+    assert_errors(visitor, [
+        UnicodeStringViolation,
+        UppercaseStringModifierViolation,
+    ])